feat: 서버 컴포넌트 page.tsx 생성, 기존 UI 로직은 ActivityPage.tsx로 분리
